requirements.txt
delta-spark==4.1.0
Faker==40.5.1
importlib_metadata==8.7.1
py4j==0.10.9.9
pyspark==4.1.1
zipp==3.23.0
